[LE_3] Test Script implementaion
