+++ LE_3/Solution/bst.cpp
@@ -133,71 +133,9 @@ public:
 
 };
 
-BST_Node* createBST(){
-    std::cout << "Level order traversal of the BST:" << std::endl;
-    BST_Node* root = new BST_Node(5);
-    root->insert(2);
-    root->insert(10);
-    root->insert(1);
-    root->insert(3);
-    root->insert(8);
-    root->insert(14);
-    root->insert(4);
-    root->insert(7);
-    root->insert(6);
-    root->insert(9);
-    root->insert(12);
-    root->insert(16);
-    root->insert(11);
-    root->insert(13);
-    root->insert(15);
-
-    root->levelOrderPrint();
-
-    return root;
-}
-
-void testFindSuccessful(BST_Node* root){
-    int keyToFind = 4;
-    bool found = root->find(keyToFind);
-    std::cout << "Key " << keyToFind << " found: " << (found ? "Yes" : "No") << std::endl;
-
-}
-
-void testFindUnsuccessful(BST_Node* root){
-    int keyToFind = 17;
-    bool found = root->find(keyToFind);
-    std::cout << "Key " << keyToFind << " found: " << (found ? "Yes" : "No") << std::endl;
-
-}
-
-void testDeleteNode(int keyToDelete, BST_Node* root){
-
-    root = root->deleteNode(keyToDelete);
-    std::cout << "After deleting key " << keyToDelete << ":" << std::endl;
-    root->levelOrderPrint();
-}
-
-void SkewedTreeComplexity(){
-    BST_Node* root = new BST_Node(1);
-    for(int i = 2; i<=10000; i++){
-        root->insert(i);
-    }
-    auto start = chrono::high_resolution_clock::now();
-    root->find(10000);
-    auto stop = chrono::high_resolution_clock::now();
-    auto duration = chrono::duration_cast<chrono::microseconds>(stop - start);
-
-    cout<<"Time taken: "<<duration.count()<<endl;
-
-}
+#ifndef TEST
 int main() {
-    // Example usage of the BST_Node class
-    BST_Node* root = createBST();
-    testFindSuccessful(root);
-    testFindUnsuccessful(root);
-    testDeleteNode(10, root);
-    SkewedTreeComplexity();
     
     return 0;
 }
+#endif
